shows active orders after submitting #12

+++ app/src/main/java/edu/hanover/hc24_luuk_crawford_senior_project/data/UserOrder.kt
@@ -7,8 +7,8 @@ import java.util.Date
  */
 data class UserOrder(
     var user: String,
-    var userID: Int,
-    val orderID: Int,
+    var userID: Long,
+    val orderID: Long,
     var itemID: Long,
     var customization: Customization,
     var itemStatus: ItemStatus,
